Complete flex assignments 01 and 02
Completed the assignments 01-flex-center and 02-flex-header
by following the READMEs and self-check list to arrive
at equivalent solutions.

Assignment 01 was completed and no difference between
my solution and the given one.

Assignment 02 was completed, and my solution worked,
but wasn't as elegant as the given solution, so more
practice is needed. I will reattempt this assignment
again and try to make a less complicated solution.