Complete flex assignment 03
Followed the README and self-check to arrive at an
equivalent solution to the assignment. My solution
had some unnecessary containers (a unique one for
each info item) and some extra centering (centered
the title and the info text rather than centering
the text for the entire body).

Commonly, with this assignment and previous, I add
too much specific information or over-complicate the
solution to the problem. This assignment I improved
over my previous but I need to continue practicing
and thinking of simpler solutions to the problems.