Complete flex assignment 05
This was the most difficult assignment for me so
far. I couldn't get the exit button to align with the
header or the margins on the text to correctly display
due to over-complicating my solution. I will revisit
this assignment after some time to see if the underlying
principles behind the solution have been properly absorbed.