Complete flex assignment 06
Followed the README and self-check to arrive at an
equivalent solution. It required some googling for
the link and list styling, but I feel like my solution
was about as elegant as the one provided. Instead
of using the flex property in the content container
I justified the body with space-between which had the
same outcome of centering the content and placing the
header and footer at the top and bottom of the page.