Complete flex assignment 07
Followed the README instructions and the self-check
items to reach an equivalent solution. It required
some rereading of the flex lessons and some research
to get the emojis to display, but the I was able to
complete it without checking the solution.

My solution had some unnecessary css properties so
my solution was not as clean as it could have been.
It took some messing around with the properties to
get the alignment correct and I should have removed
the unneeded properties after getting everything
set prior to checking the solution. It was a good
learning experience.